# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMFACA Crédito API",
        "description": "API para la gestión de solicitudes de crédito de COMFACA",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/lineas-inversion/{linea_id}": {
            "get": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Obtener línea de inversión por ID",
                "description": "Obtener una línea de inversión específica por su ID.\n\nArgs:\n    * @param int $linea_id ID de la línea de inversión\nReturns:\n    * JsonResponse\nResponse:\n    * {\n        \"success\": true,\n        \"message\": \"Línea de inversión obtenida exitosamente\",\n        \"data\": {...}\n    }",
                "operationId": "6d042142c4a56d8f137fd4e8b3be5d0e",
                "parameters": [
                    {
                        "name": "linea_id",
                        "in": "path",
                        "description": "ID de la línea",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Línea obtenida"
                    },
                    "404": {
                        "description": "Línea no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-inversion/categoria/{categoria}": {
            "get": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Obtener líneas por categoría",
                "description": "Obtener líneas de inversión por categoría.\n\nArgs:\n    * @param string $categoria Categoría de la línea (EDUCACION, VIVIENDA, etc.)\nReturns:\n    * JsonResponse\nResponse:\n    * {\n        \"success\": true,\n        \"message\": \"Líneas de inversión por categoría obtenidas\",\n        \"data\": [...]\n    }",
                "operationId": "deda1b1c0690d440dd9014eab8b766e6",
                "parameters": [
                    {
                        "name": "categoria",
                        "in": "path",
                        "description": "Categoría",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Líneas por categoría obtenidas"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-inversion/inicializar": {
            "post": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Inicializar líneas de inversión",
                "description": "Inicializar líneas de inversión",
                "operationId": "8f84faa327cec44c719d7402c97b5a5d",
                "responses": {
                    "200": {
                        "description": "Líneas inicializadas"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-inversion/estadisticas": {
            "get": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Obtener estadísticas de líneas de inversión",
                "description": "Obtener estadísticas de líneas de inversión",
                "operationId": "970968985d5e918415d6707c968dfbbc",
                "responses": {
                    "200": {
                        "description": "Estadísticas obtenidas"
                    },
                    "401": {
                        "description": "No autorizado"
                    },
                    "500": {
                        "description": "Error del servidor"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-inversion/buscar": {
            "get": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Buscar líneas de inversión",
                "description": "Buscar líneas de inversión por texto",
                "operationId": "0280664d5f850063443a4172d8734af0",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Texto de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categoria",
                        "in": "query",
                        "description": "Filtrar por categoría",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    },
                    "422": {
                        "description": "Parámetros inválidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/lineas-inversion/comparar": {
            "post": {
                "tags": [
                    "LineasInversion"
                ],
                "summary": "Comparar líneas de inversión",
                "description": "Comparar líneas de inversión",
                "operationId": "8e0f17e251b5edf9977591c089bdf348",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "lineas"
                                ],
                                "properties": {
                                    "lineas": {
                                        "description": "Array de IDs de líneas a comparar",
                                        "type": "array",
                                        "items": {
                                            "type": "integer"
                                        },
                                        "example": [
                                            1,
                                            2,
                                            3
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Comparación completada"
                    },
                    "422": {
                        "description": "Parámetros inválidos"
                    },
                    "401": {
                        "description": "No autorizado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
}